feat: forgot password section code modular done
