+++ src/index.html
@@ -301,6 +301,6 @@
     ></div>
 
     <script src="/js/main.js" type="module"></script>
-    <script type="module" src="/js/profile.js"></script>
+    <script type="module" src="/js/profile/profile.js"></script>
   </body>
 </html>
