+++ src/js/pages/api/AuthApi.js
@@ -1,5 +1,5 @@
 import axios from "axios";
-const API_BASE_URL = "http://localhost:3001/api/auth";
+const API_BASE_URL = "http://localhost:3001/user/auth";
 
 export default class AuthApi {
   api = axios.create({
