+++ src/js/pages/forgotPassword/forgotPassword.js
@@ -1,52 +1,10 @@
 
-// Import all of Bootstrap’s JS
-import * as bootstrap from "bootstrap";
-import { showToast } from "../../showToast.js";
-import authApi from "../api/AuthApi.js";
-const authAPI = new authApi();
+// const forgotPassForm = document.querySelector('form');
+// const emailInput = document.getElementById('email-input');
+// const submitButton = document.querySelector('button[type="submit"]');
+import { forgotPasswordPage } from "../../dom/domHandler.js";
+import { handleForgotPass } from "./eventHandlers.js";
 
-const forgotPassForm = document.querySelector('form');
-const emailInput = document.getElementById('email-input');
-const submitButton = document.querySelector('button[type="submit"]');
-
-async function handleForgotPass(event) {
-  event.preventDefault();
-
-  const email = emailInput.value.trim();
-
-
-  const emailRegex = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
-  if (!emailRegex.test(email)) {
-    showToast('Please enter a valid email address', "error");
-    return;
-  }
-
-  const originalText = submitButton.textContent;
-  submitButton.textContent = 'Sending otp...';
-  submitButton.disabled = true;
-
-  try {
-    const response1 = await authAPI.forgetPasswordSendOtp(email);
-    // console.log(response1);
-
-    localStorage.setItem('reset_email', email);
-
-    showToast('Otp sent! Check your email...', "success");
-
-    setTimeout(() => {
-      window.location.href = '../../pages/forgotOtpVerify.html';
-    }, 1000);
-
-
-  } catch (error) {
-    showToast(error.message, "error" || 'Otp sending failed. Please try again...', "error");
-  } finally {
-    submitButton.textContent = originalText;
-    submitButton.disabled = false;
-  }
+if (forgotPasswordPage.forgotPassForm) {
+  forgotPasswordPage.forgotPassForm.addEventListener('submit', handleForgotPass);
 }
-
-if (forgotPassForm) {
-  forgotPassForm.addEventListener('submit', handleForgotPass);
-}
-
