--- src/js/profile.js
@@ -1,7 +1,0 @@
-// import '../scss/styles.scss';
-// eslint-disable-next-line 
-import * as bootstrap from 'bootstrap';
-
-import { profileIcon } from './dom.js';
-
-profileIcon();
