+++ src/js/profile/profile.js
@@ -1,0 +1,4 @@
+
+import { profileIcon } from './profileHandler';
+
+profileIcon();
